Move filesystem.usb package to file_operations module
