+++ file_operations/src/main/java/com/amaze/filemanager/file_operations/filesystem/usb/UsbOtgRepresentation.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.filesystem.usb;
+package com.amaze.filemanager.file_operations.filesystem.usb;
 
 import androidx.annotation.Nullable;
